Moved sprite classes to another file
